refactor(photo-section): separate server and client components
