+++ src/widgets/photo-grid/index.ts
@@ -1,0 +1,1 @@
+export { PhotoGrid } from './ui/PhotoGrid';
